Modified ready-for-grading logic
Previously, any submission that was ready for grading was
treated as if it had been sent to the graders (even if
an instructor had not yet created the master grading
repos). We now explicitly distinguish between "ready
for grading" and "sent to the graders". This means
that students can cancel and re-submit submissions
up until their code has actually been sent to the
graders (which will always require manual intervention
from an instructor)